Update 104-fibonacci.c file

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,7 +8,7 @@
 int main(void)
 {
 	int count;
-	unsigned long long num1 = 1, num2 = 2, next;
+	unsigned long num1 = 1, num2 = 2, next;
 
 	printf("%llu, %llu, ", num1, num2);
